# docs/openapi.yaml
openapi: 3.0.3
info:
  title: TorahTrain Backend API
  version: 1.0.0
servers:
  - url: /
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    CreateStudentRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        displayName:
          type: string
    UpdateCoinsRequest:
      type: object
      required: [delta]
      properties:
        delta:
          type: integer
    FeedbackRequest:
      type: object
      required: [message]
      properties:
        message:
          type: string
        score:
          type: number
    AssignmentRequest:
      type: object
      required: [studentId]
      properties:
        studentId:
          type: integer
        startVerseId:
          type: integer
        endVerseId:
          type: integer
        partStartIndex:
          type: integer
        partEndIndex:
          type: integer
paths:
  /api/auth/login:
    post:
      summary: Log in and receive JWT
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated
        '401':
          description: Invalid credentials
  /api/texts/{bookId}/{chapter}:
    get:
      summary: Fetch text hierarchy for a chapter
      tags: [Texts]
      security:
        - bearerAuth: []
      parameters:
        - name: bookId
          in: path
          required: true
          schema: { type: integer }
        - name: chapter
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Hierarchical verses, words, and tropes
  /api/students:
    post:
      summary: Create student profile (ADMIN only)
      tags: [Students]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStudentRequest'
      responses:
        '201':
          description: Student created
    get:
      summary: List student profiles (ADMIN only)
      tags: [Students]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Student list
  /api/students/{id}:
    get:
      summary: Get student profile (ADMIN only)
      tags: [Students]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Student profile
    put:
      summary: Update student profile (ADMIN only)
      tags: [Students]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Updated profile
    delete:
      summary: Delete student profile (ADMIN only)
      tags: [Students]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '204':
          description: Deleted
  /api/students/{id}/coins:
    post:
      summary: Update student coins (ADMIN only)
      tags: [Gamification]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCoinsRequest'
      responses:
        '200':
          description: Coin balance updated
  /api/students/{id}/feedback:
    post:
      summary: Add student feedback (ADMIN only)
      tags: [Gamification]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedbackRequest'
      responses:
        '200':
          description: Feedback stored
  /api/assignments:
    post:
      summary: Create assignment route (ADMIN only)
      tags: [Assignments]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignmentRequest'
      responses:
        '201':
          description: Assignment created
  /api/assignments/my-routes:
    get:
      summary: Get active assignments for current student
      tags: [Assignments]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Active assignment routes